Added save product

+++ application/src/main/java/com/mlucov/business/product/save/SaveProductUseCaseApi.java
@@ -3,6 +3,6 @@ package com.mlucov.business.product.save;
 import com.mlucov.business.product.models.SaveProductInput;
 
 public interface SaveProductUseCaseApi {
-    Integer createProduct(SaveProductInput input);
-    Integer updateProduct(SaveProductInput input, Integer id);
+    Long createProduct(SaveProductInput input);
+    Long updateProduct(SaveProductInput input, Long id);
 }
